+++ src/components/textarea/TextArea.js
@@ -1,7 +1,8 @@
-/* global ace */
+/* global ace, Quill */
 import TextFieldComponent from '../textfield/TextField';
 import Formio from '../../Formio';
 import _ from 'lodash';
+import { uniqueName } from '../../utils/utils';
 
 export default class TextAreaComponent extends TextFieldComponent {
   static schema(...extend) {
@@ -74,6 +75,7 @@ export default class TextAreaComponent extends TextFieldComponent {
   }
 
   createInput(container) {
+    const _this = this;
     if (this.isPlain) {
       if (this.options.readOnly) {
         this.input = this.ce('div', {
@@ -160,6 +162,9 @@ export default class TextAreaComponent extends TextFieldComponent {
         this.updateValue(null, this.getConvertedValue(this.quill.root.innerHTML));
       }
     ).then((quill) => {
+      if (this.component.isUploadEnabled) {
+        quill.getModule('toolbar').addHandler('image', imageHandler);
+      }
       quill.root.spellcheck = this.component.spellcheck;
       if (this.options.readOnly || this.component.disabled) {
         quill.disable();
@@ -169,6 +174,58 @@ export default class TextAreaComponent extends TextFieldComponent {
     }).catch(err => console.warn(err));
 
     return this.input;
+
+    function imageHandler() {
+      let fileInput = this.container.querySelector('input.ql-image[type=file]');
+
+      if (fileInput == null) {
+        fileInput = document.createElement('input');
+        fileInput.setAttribute('type', 'file');
+        fileInput.setAttribute('accept', 'image/*');
+        fileInput.classList.add('ql-image');
+        fileInput.addEventListener('change', () => {
+          const files = fileInput.files;
+          const range = this.quill.getSelection(true);
+
+          if (!files || !files.length) {
+            console.warn('No files selected');
+            return;
+          }
+
+          this.quill.enable(false);
+          const { uploadStorage, uploadUrl, uploadOptions, uploadDir } = _this.component;
+          _this.root.formio
+            .uploadFile(
+              uploadStorage,
+              files[0],
+              uniqueName(files[0].name),
+              uploadDir || '', //should pass empty string if undefined
+              null,
+              uploadUrl,
+              uploadOptions
+            )
+            .then(result => {
+              return _this.root.formio.downloadFile(result);
+            })
+            .then(result => {
+              this.quill.enable(true);
+              const Delta = Quill.import('delta');
+              this.quill.updateContents(new Delta()
+                  .retain(range.index)
+                  .delete(range.length)
+                  .insert({ image: result.url })
+                , Quill.sources.USER);
+              fileInput.value = '';
+            }).catch(error => {
+            console.warn('Quill image upload failed');
+            console.warn(error);
+            this.quill.enable(true);
+          });
+        });
+        this.container.appendChild(fileInput);
+      }
+      fileInput.click();
+    }
   }
 
   setConvertedValue(value) {